requirements.txt
launchdarkly-server-sdk
python-dotenv
resend